Move SendDataResult and RetryStrategy for trace_utils::SendDataResult
to separate files and add rustdoc commments for public functions.
