+++ trace-utils/src/trace_utils.rs
@@ -7,11 +7,11 @@ use std::cmp::Ordering;
 use std::collections::HashMap;
 
 pub use crate::send_data::SendData;
-pub use crate::send_data::SendDataResult;
+// TODO: EK - FIX THIS
+pub use crate::send_data::send_data_result::SendDataResult;
 pub use crate::tracer_header_tags::TracerHeaderTags;
 use datadog_trace_normalization::normalizer;
-use datadog_trace_protobuf::pb;
-use datadog_trace_protobuf::pb::TraceChunk;
+use datadog_trace_protobuf::pb::{self, Span, TraceChunk, TracerPayload};
 use ddcommon::azure_app_services;
 
 /// Span metric the mini agent must set for the backend to recognize top level span
@@ -22,13 +22,11 @@ const TRACER_TOP_LEVEL_KEY: &str = "_dd.top_level";
 const MAX_PAYLOAD_SIZE: usize = 50 * 1024 * 1024;
 
 /// First value of returned tuple is the payload size
-pub async fn get_traces_from_request_body(
-    body: Body,
-) -> anyhow::Result<(usize, Vec<Vec<pb::Span>>)> {
+pub async fn get_traces_from_request_body(body: Body) -> anyhow::Result<(usize, Vec<Vec<Span>>)> {
     let buffer = hyper::body::aggregate(body).await?;
     let size = buffer.remaining();
 
-    let traces: Vec<Vec<pb::Span>> = match rmp_serde::from_read(buffer.reader()) {
+    let traces: Vec<Vec<Span>> = match rmp_serde::from_read(buffer.reader()) {
         Ok(res) => res,
         Err(err) => {
             anyhow::bail!("Error deserializing trace from request body: {err}")
@@ -51,8 +49,8 @@ pub struct RootSpanTags<'a> {
     pub runtime_id: &'a str,
 }
 
-pub(crate) fn construct_trace_chunk(trace: Vec<pb::Span>) -> pb::TraceChunk {
-    pb::TraceChunk {
+pub(crate) fn construct_trace_chunk(trace: Vec<Span>) -> TraceChunk {
+    TraceChunk {
         priority: normalizer::SamplerPriority::None as i32,
         origin: "".to_string(),
         spans: trace,
@@ -62,11 +60,11 @@ pub(crate) fn construct_trace_chunk(trace: Vec<pb::Span>) -> pb::TraceChunk {
 }
 
 pub(crate) fn construct_tracer_payload(
-    chunks: Vec<pb::TraceChunk>,
+    chunks: Vec<TraceChunk>,
     tracer_tags: &TracerHeaderTags,
     root_span_tags: RootSpanTags,
-) -> pb::TracerPayload {
-    pb::TracerPayload {
+) -> TracerPayload {
+    TracerPayload {
         app_version: root_span_tags.app_version.to_string(),
         language_name: tracer_tags.lang.to_string(),
         container_id: tracer_tags.container_id.to_string(),
@@ -134,13 +132,13 @@ pub fn coalesce_send_data(mut data: Vec<SendData>) -> Vec<SendData> {
     data
 }
 
-fn get_root_span_index(trace: &Vec<pb::Span>) -> anyhow::Result<usize> {
+fn get_root_span_index(trace: &Vec<Span>) -> anyhow::Result<usize> {
     if trace.is_empty() {
         anyhow::bail!("Cannot find root span index in an empty trace.");
     }
 
     // parent_id -> (child_span, index_of_child_span_in_trace)
-    let mut parent_id_to_child_map: HashMap<u64, (&pb::Span, usize)> = HashMap::new();
+    let mut parent_id_to_child_map: HashMap<u64, (&Span, usize)> = HashMap::new();
 
     // look for the span with parent_id == 0 (starting from the end) since some clients put the root
     // span last.
@@ -185,7 +183,7 @@ fn get_root_span_index(trace: &Vec<pb::Span>) -> anyhow::Result<usize> {
 ///   - OR its parent is unknown (other part of the code, distributed trace)
 ///   - OR its parent belongs to another service (in that case it's a "local root" being the highest
 ///     ancestor of other spans belonging to this service and attached to it).
-pub fn compute_top_level_span(trace: &mut [pb::Span]) {
+pub fn compute_top_level_span(trace: &mut [Span]) {
     let mut span_id_to_service: HashMap<u64, String> = HashMap::new();
     for span in trace.iter() {
         span_id_to_service.insert(span.span_id, span.service.clone());
@@ -210,7 +208,7 @@ pub fn compute_top_level_span(trace: &mut [pb::Span]) {
     }
 }
 
-fn set_top_level_span(span: &mut pb::Span, is_top_level: bool) {
+fn set_top_level_span(span: &mut Span, is_top_level: bool) {
     if !is_top_level {
         if span.metrics.contains_key(TOP_LEVEL_KEY) {
             span.metrics.remove(TOP_LEVEL_KEY);
@@ -221,7 +219,7 @@ fn set_top_level_span(span: &mut pb::Span, is_top_level: bool) {
 }
 
 pub fn set_serverless_root_span_tags(
-    span: &mut pb::Span,
+    span: &mut Span,
     function_name: Option<String>,
     env_type: &EnvironmentType,
 ) {
@@ -240,7 +238,7 @@ pub fn set_serverless_root_span_tags(
     }
 }
 
-fn update_tracer_top_level(span: &mut pb::Span) {
+fn update_tracer_top_level(span: &mut Span) {
     if span.metrics.contains_key(TRACER_TOP_LEVEL_KEY) {
         span.metrics.insert(TOP_LEVEL_KEY.to_string(), 1.0);
     }
@@ -259,7 +257,7 @@ pub struct MiniAgentMetadata {
 }
 
 pub fn enrich_span_with_mini_agent_metadata(
-    span: &mut pb::Span,
+    span: &mut Span,
     mini_agent_metadata: &MiniAgentMetadata,
 ) {
     if let Some(gcp_project_id) = &mini_agent_metadata.gcp_project_id {
@@ -272,7 +270,7 @@ pub fn enrich_span_with_mini_agent_metadata(
     }
 }
 
-pub fn enrich_span_with_azure_metadata(span: &mut pb::Span, mini_agent_version: &str) {
+pub fn enrich_span_with_azure_metadata(span: &mut Span, mini_agent_version: &str) {
     if let Some(aas_metadata) = azure_app_services::get_function_metadata() {
         let aas_tags = [
             ("aas.resource.id", aas_metadata.get_resource_id()),
@@ -313,12 +311,12 @@ macro_rules! parse_root_span_tags {
 }
 
 pub fn collect_trace_chunks(
-    mut traces: Vec<Vec<pb::Span>>,
+    mut traces: Vec<Vec<Span>>,
     tracer_header_tags: &TracerHeaderTags,
     process_chunk: impl Fn(&mut TraceChunk, usize),
     is_agentless: bool,
-) -> pb::TracerPayload {
-    let mut trace_chunks: Vec<pb::TraceChunk> = Vec::new();
+) -> TracerPayload {
+    let mut trace_chunks: Vec<TraceChunk> = Vec::new();
 
     // We'll skip setting the global metadata and rely on the agent to unpack these
     let mut gathered_root_span_tags = !is_agentless;
@@ -390,20 +388,21 @@ mod tests {
         test_utils::create_test_span,
         trace_utils::{self, SendData},
     };
-    use datadog_trace_protobuf::pb;
+    use datadog_trace_protobuf::pb::TraceChunk;
+    use datadog_trace_protobuf::pb::{Span, TracerPayload};
     use ddcommon::Endpoint;
 
     #[test]
     fn test_coalescing_does_not_exceed_max_size() {
         let dummy = SendData::new(
             MAX_PAYLOAD_SIZE / 5 + 1,
-            pb::TracerPayload {
+            TracerPayload {
                 container_id: "".to_string(),
                 language_name: "".to_string(),
                 language_version: "".to_string(),
                 tracer_version: "".to_string(),
                 runtime_id: "".to_string(),
-                chunks: vec![pb::TraceChunk {
+                chunks: vec![TraceChunk {
                     priority: 0,
                     origin: "".to_string(),
                     spans: vec![],
@@ -471,7 +470,7 @@ mod tests {
                     "meta": {},
                     "metrics": {},
                 }]),
-                vec![vec![pb::Span {
+                vec![vec![Span {
                     service: "test-service".to_string(),
                     name: "test-service-name".to_string(),
                     resource: "test-service-resource".to_string(),
@@ -498,7 +497,7 @@ mod tests {
                     "duration": 5,
                     "meta": {},
                 }]),
-                vec![vec![pb::Span {
+                vec![vec![Span {
                     service: "".to_string(),
                     name: "test-service-name".to_string(),
                     resource: "test-service-resource".to_string(),
